Factor out SEP from main
