+++ .idea/workspace.xml
@@ -16,12 +16,11 @@
   </component>
   <component name="ChangeListManager">
     <list default="true" id="72450d24-770a-4879-b33d-e011e8325810" name="Changes" comment="">
-      <change afterPath="$PROJECT_DIR$/src/lzvn.rs" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/Cargo.toml" beforeDir="false" afterPath="$PROJECT_DIR$/Cargo.toml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/build.rs" beforeDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/ext/lzvn" beforeDir="false" />
+      <change afterPath="$PROJECT_DIR$/src/sep.rs" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/tests/fixtures/sep.dec" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/tests/integration_tests.rs" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/src/main.rs" beforeDir="false" afterPath="$PROJECT_DIR$/src/main.rs" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/wrapper.h" beforeDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -61,13 +60,18 @@
     "WebServerToolWindowFactoryState": "false",
     "cf.first.check.clang-format": "false",
     "cidr.known.project.marker": "true",
-    "last_opened_file_path": "/Users/rickmark/Developer/sepsplit-rs",
+    "last_opened_file_path": "/Users/rickmark/Developer/sepsplit-rs/tests/fixtures",
     "nodejs_package_manager_path": "npm",
     "org.rust.cargo.project.model.PROJECT_DISCOVERY": "true",
     "settings.editor.selected.configurable": "language.rust.build.tool.cargo"
   }
 }]]></component>
-  <component name="RunManager" selected="Cargo.Run">
+  <component name="RecentsManager">
+    <key name="CopyFile.RECENT_KEYS">
+      <recent name="$PROJECT_DIR$/tests/fixtures" />
+    </key>
+  </component>
+  <component name="RunManager" selected="Cargo.Build">
     <configuration name="Build" type="CargoCommandRunConfiguration" factoryName="Cargo Command">
       <option name="command" value="build" />
       <option name="workingDirectory" value="file://$PROJECT_DIR$" />
@@ -102,7 +106,25 @@
         <option name="CARGO.BUILD_TASK_PROVIDER" enabled="true" />
       </method>
     </configuration>
+    <configuration name="Test" type="CargoCommandRunConfiguration" factoryName="Cargo Command">
+      <option name="command" value="test" />
+      <option name="workingDirectory" value="file://$PROJECT_DIR$" />
+      <option name="channel" value="NIGHTLY" />
+      <option name="requiredFeatures" value="true" />
+      <option name="allFeatures" value="false" />
+      <option name="emulateTerminal" value="false" />
+      <option name="withSudo" value="false" />
+      <option name="buildTarget" value="REMOTE" />
+      <option name="backtrace" value="SHORT" />
+      <envs />
+      <option name="isRedirectInput" value="false" />
+      <option name="redirectInputPath" value="" />
+      <method v="2">
+        <option name="CARGO.BUILD_TASK_PROVIDER" enabled="true" />
+      </method>
+    </configuration>
     <list>
+      <item itemvalue="Cargo.Test" />
       <item itemvalue="Cargo.Build" />
       <item itemvalue="Cargo.Run" />
     </list>
@@ -120,7 +142,7 @@
       <option name="presentableId" value="Default" />
       <updated>1662693504261</updated>
       <workItem from="1662693505694" duration="4113000" />
-      <workItem from="1662750422273" duration="2877000" />
+      <workItem from="1662750422273" duration="3891000" />
     </task>
     <servers />
   </component>
